+++ src/PairList.h
@@ -1,5 +1,5 @@
-#ifndef INC_PAIRLIST_H
-#define INC_PAIRLIST_H
+#ifndef INC_PAIRLIST2_H
+#define INC_PAIRLIST2_H
 #include "Topology.h"
 #include "Timer.h"
 /// Class for creating a lists of potential pairing atoms via spatial grid.
@@ -8,12 +8,17 @@
   *       However, it is more memory-hungry since it stores full cell
   *       neighbor lists instead of cell centers.
   */
-class PairList {
+class PairList2 {
   public:
     typedef std::vector<int> Iarray;
-    typedef std::vector<Iarray> Iarray2D;
-    typedef std::vector<Vec3> Varray;
-    PairList();
+    /// PairList Atom
+    class AtmType;
+    typedef std::vector<AtmType> Aarray;
+    /// PairList Cell
+    class CellType;
+    typedef std::vector<CellType> Carray;
+
+    PairList2();
     /// Initialize pair list with given cutoff, "skin", and debug level.
     int InitPairList(double,double,int);
     /// Setup pair list grid cells based on given box and vector of recip lengths.
@@ -24,68 +29,88 @@ class PairList {
     void Timing(double) const;
     /// Print memory usage.
     void PrintMemory() const;
-    /// \return Total number of grid cells.
-    int NGridMax()                 const { return nGridMax_;         }
-    /// \return Array of neighbor grid cell indices for specified grid cell.
-    Iarray const& Cell(int i)      const { return neighborPtr_[i];   }
-    /// \return Array of indices into TransVec for each neighbor grid cell in Cell(i).
-    Iarray const& Trans(int i)     const { return neighborTrans_[i]; }
-    /// \return Starting index into AtomGridIdx for given grid cell.
-    int IdxOffset(int i)           const { return idxOffset_[i];     }
-    /// \return Number of atoms in given grid cell.
-    int NatomsInGrid(int i)        const { return nAtomsInGrid_[i];  }
-    /// \return Atom index for given index (related to IdxOffset)
-    int AtomGridIdx(int a)         const { return atomGridIdx_[a];   }
-    /// \return Imaged coordinates for given atom index.
-    Vec3 const& ImageCoords(int n) const { return Image_[n];         }
-    /// \return Translation vector for given translation index (from Trans()).
+    /// \return Number of grid cells.
+    int NGridMax()                const { return (int)cells_.size(); }
+    /// \return Specified grid cell.
+    CellType const& Cell(int idx) const { return cells_[idx];        }
+    /// \return Translation vector for given translation index (from TransList()).
     Vec3 const& TransVec(int t)    const { return translateVec_[t];  }
-
-    Varray const& ImageCoords()    const { return Image_; }
-    Varray const& FracCoords()     const { return Frac_;  }
   private:
-//    int Fill_CellNeighbor();
     /// Determine neighbors and translation vectors for each cell.
     void CalcGridPointers(int,int);
     /// Check grid dimensions using given recip lengths and (re)allocate mem if necessary.
     int SetupGrids(Vec3 const&);
-    /// Convert selected coords to wrapped fraction coords and save wrapped Cartesian coords.
-    void MapCoords(Frame const&, Matrix_3x3 const&,Matrix_3x3 const&, AtomMask const&);
     /// Update the translation vectors based on given unit cell matrix.
     void FillTranslateVec(Matrix_3x3 const&);
-    /// Assign fraction coords to grid cells.
-    void GridUnitCell();
+    /// Add atoms to grid cells.
+    void GridUnitCell(Frame const&, Matrix_3x3 const&, Matrix_3x3 const&, AtomMask const&);
+    /// Calculate cell index based on given coords and add to cell
+    inline void GridAtom(int, Vec3 const&, Vec3 const&);
 
-//    typedef std::vector<bool> Barray;
-
-    static const int cellOffset_; ///< Number of cells in forward direction to check
-
+    Carray cells_;            ///< Hold all cells in grid
     Vec3 translateVec_[18];   ///< Translate vector array
-//    int cellNeighbor_[7][10]; ///< Cell neighbor index array based on cellOffset_.
-    Iarray2D neighborPtr_;    ///< For each grid cell, hold indexes into neighbor cells.
-    Iarray2D neighborTrans_;
-    Varray Frac_;             ///< Hold fractional coords back in primary cell.
-    Varray Image_;            ///< Hold Cartesian coords back in primary cell.
-//    Iarray nLoGrid_;
-//    Iarray nHiGrid_;
-    Iarray nAtomsInGrid_;     ///< Number of atoms in each grid cell.
-    Iarray idxOffset_;        ///< Offset of starting atom in each grid cell.
-    Iarray atomCell_;         ///< Grid cell index for each atom.
-    Iarray atomGridIdx_;      ///< List of atoms sorted by grid cell.
-//    Barray myGrids_;          ///< True if I am responsible for this grid cell.
     double cutList_;          ///< Direct space cutoff plus non-bond "skin"
     int debug_;
-    int NP_;                  ///< Total number of elements in neighborPtr_/neighborTrans_
     int nGridX_;              ///< Number of grid cells in X direction.
     int nGridY_;              ///< Number of grid cells in Y direction.
     int nGridZ_;              ///< Number of grid cells in Z direction.
-    int nGridMax_;            ///< Total number of grid cells.
     int nGridX_0_;            ///< Previous number of cells in X direction
     int nGridY_0_;            ///< Previous number of cells in Y direction
     int nGridZ_0_;            ///< Previous number of cells in Z direction
-//    int maxNptrs_;            ///< Max number of neighbor pointers.
+    static const int cellOffset_; ///< Number of cells in forward direction to check
+
     Timer t_map_;
     Timer t_gridpointers_;
     Timer t_total_;
+};
+// -----------------------------------------------------------------------------
+/** PairList Atom. Holds atom index and wrapped coordinates. */
+class PairList2::AtmType {
+  public:
+    AtmType() : idx_(-1) {}
+    /// CONSTRUCTOR - Atom index, wrapped fractional coords, wrapped Cart. coords
+    AtmType(int i, Vec3 const& f, Vec3 const& c) :
+      imageCoords_(c), fracCoords_(f), idx_(i) {}
+    /// \return Wrapped Cart. coords
+    Vec3 const& ImageCoords() const { return imageCoords_; }
+    /// \return Atom index.
+    int Idx()                 const { return idx_;         }
+  private:
+    Vec3 imageCoords_; ///< Imaged Cartesian coordinates
+    Vec3 fracCoords_;  ///< Fractional coordinates
+    int idx_;          ///< Atom index
+};
+// -----------------------------------------------------------------------------
+/** PairList Cell. Holds list of atoms and neighbor cells. */
+class PairList2::CellType {
+  public:
+    CellType() {}
+    /// Add given atom to this cell
+    void AddAtom(AtmType const& a) { atoms_.push_back( a ); }
+    /// Clear all atoms in this cell
+    void ClearAtoms()              { atoms_.clear();        }
+    /// \return Number of atoms in the cell
+    unsigned int NatomsInGrid() const { return atoms_.size();  }
+    /// \return List of cell neighbors - starts with this cell
+    Iarray const& CellList()    const { return neighborPtr_;   }
+    /// \return List of translations for neighbors - starts with this cell
+    Iarray const& TransList()   const { return neighborTrans_; }
+    /// \return Memory used by this cell in bytes
+    size_t MemSize() const {
+      return ((neighborPtr_.size() + neighborTrans_.size()) * sizeof(int)) +
+             (2 * sizeof(Iarray)) + (atoms_.size() * sizeof(AtmType)) + sizeof(Aarray);
+    }
+    /// Atom list iterator
+    typedef Aarray::const_iterator const_iterator;
+    /// \return Iterator to beginning of atom list.
+    const_iterator begin() const { return atoms_.begin(); }
+    /// \return Iterator to end of atom list.
+    const_iterator end()   const { return atoms_.end();   }
+    // PairList2 is a friend so it can access neighborPtr_ and neighborTrans_
+    friend class PairList2;
+  private:
+    Iarray neighborPtr_;   ///< Indices of neighbor cells "forward" of this cell.
+    Iarray neighborTrans_; ///< Index pointing to translate vector for each neighbor cell.
+    Aarray atoms_;         ///< Atoms in this cell.
 };
 #endif
